Updates ShareAgent deployment and monitoring
Increases the cycle allocation for ShareAgent deployment to ensure successful creation.

Adds a command to update the limit for creating ShareAgent mAIners in the game state canister.

Reduces the polling delay in the gamestate log monitor for quicker updates.

Adds a new file containing the canister IDs of the mainers on the prd network.