Admin RBAC for mAIners
Updates the admin RBAC scripts to use record syntax for assigning roles.

Adds a `--deploy-with-yes` option to the upgrade scripts to skip confirmation prompts during deployment. This enhances automation.

Also creates a logs folder to capture the output of the admin RBAC script.

Removes logs for admin RBAC update

Adds *.logs to .gitignore